// types/bookStand.ts
export interface BookStand {
  id: string;
  referenceId: string;
  companyName: string;
  contactName: string;
  email: string;
  phone: string;
  standType: string;
  paymentMethod: string;
  additionalRequirements: string;
  status: string;
  adminConfirmed: boolean;
  createdAt: string;
  updatedAt: string;
} 